\name{concatenate}
\alias{concatenate}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{To concatenate matrices columnwise}
\description{
   This is an internal function.
           
}
\usage{
 concatenate(datasets)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{datasets}{A list of matrices}

}

\details{
        An internal function for package.
}
\value{
     Concatenated matirx.
  
}

\author{Abhishek Tripathi, Arto Klami}


\examples{

     #data(expdata1)
     #data(expdata2)

     #concatenate(list(expdata1,expdata2))

}

\keyword{internal}% at least one, from doc/KEYWORDS

