% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_drum_matrix}
\alias{demo_drum_matrix}
\title{DMT demo/instruction drum matrix}
\description{
The drum patterns used for the practice trials before the main test,
in order of use (\code{TrialNo}): Easy_1 (kick on 1 and 3), Practice_2
(kick on 1, snare on 2 and 4, hi-hat on 3), Easy_2 (snare), Easy_3
(hi-hat) and Easy_4 (kick and snare).
}
