\name{dependency.score}
\alias{dependency.score}
\title{dependency.score}
\description{Dependency score}
\usage{
dependency.score(model)
}
\arguments{
  \item{model}{DependencyModel object}
}
\value{Dependency score}
\references{See citation("dmt").}
\author{Leo Lahti \email{leo.lahti@iki.fi}}
\examples{
#dependency.score(model)
}
\keyword{ utilities }
\keyword{ maths}

