% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMT.R
\name{DMT_languages}
\alias{DMT_languages}
\title{DMT languages}
\usage{
DMT_languages()
}
\value{
A character vector of language codes.
}
\description{
Lists the languages available for the DMT. Each language corresponds to a
column in \code{data-raw/DMT_dict.xlsx}.
}
