% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drum_matrix}
\alias{drum_matrix}
\title{DMT drum matrix (normal difficulty)}
\description{
The item bank of "normal"-difficulty drum patterns used in the main
trials of the Drum Machine Test. Each row represents one onset
(Instrument x BeatPositionSixteenth) belonging to a given Stimulus.
}
