% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{easy_stimuli_drum_matrix}
\alias{easy_stimuli_drum_matrix}
\title{DMT easy stimuli drum matrix}
\description{
The item bank of "easy"-difficulty drum patterns, used alongside
\code{\link{drum_matrix}} for stratified sampling of main trials.
}
