% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{data}
\name{expdata}
\alias{expdata}
\alias{expdata1}
\alias{expdata2}
\title{Example data for drCCA}
\format{A data matrix with 2000 rows and 6 columns. Column names and row
names are ommitted.}
\description{
Randomly generated data set for the examples.  Data matrix with 2000 rows
and 6 columns.  Rows corresponds to the samples and each column corresponds
to a particular features.
}
\details{
A randomly generated data set for the example purpose only.
}
\examples{
#data(expdata1)
}
\keyword{datasets}

