% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_summary.R
\name{lm_summary}
\alias{lm_summary}
\title{Linear Models Summary Table}
\usage{
lm_summary(
  model_data,
  Sigma = FALSE,
  AIC = TRUE,
  BIC = TRUE,
  LogLik = TRUE,
  RMSE = TRUE
)
}
\arguments{
\item{model_data}{An lm() object from which the data is generated.}

\item{Sigma}{A boolean value that controls the display of the Sigma variable.}

\item{AIC}{A boolean value that controls the display of the AIC variable.}

\item{BIC}{A boolean value that controls the display of the BIC variable.}

\item{LogLik}{A boolean value that controls the display of the LogLik variable.}

\item{RMSE}{A boolean value that controls the display of the RMSE variable.}
}
\value{
Az összeg
}
\description{
This function creates a model summary table from one or more lm objects.
}
\examples{
model_data <- lm(dist ~ speed, data = cars)
lm_summary(model_data)
}
