% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.R
\name{pfa}
\alias{pfa}
\title{Probabilistic Factor Analysis}
\usage{
pfa(X, Y = NULL, zDimension = NULL, includeData = TRUE,
  calculateZ = TRUE, priors = NULL)
}
\arguments{
\item{X}{X}

\item{Y}{Y}

\item{zDimension}{zDimension}

\item{includeData}{includeData}

\item{calculateZ}{calculateZ}

\item{priors}{priors}
}
\value{
pfa model
}
\description{
Probabilistic factor analysis.
}
\examples{
#
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("dmt").
}
\keyword{utilities}

