% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppca.R
\name{ppca}
\alias{ppca}
\title{Probabilistic PCA}
\usage{
ppca(X, Y = NULL, zDimension = NULL, includeData = TRUE,
  calculateZ = TRUE)
}
\arguments{
\item{X}{Data set (variables x samples)}

\item{Y}{The second data set is optional.}

\item{zDimension}{Dimensionality of the shared latent variable.}

\item{includeData}{Logical indicating whether the original data is included
to the model output. Using \code{FALSE} can be used to save memory.}

\item{calculateZ}{Logical indicating whether an expectation of the latent
variable Z is included in the model output. Otherwise the expectation can be
calculated with \code{getZ} or \code{z.expectation}. Using \code{FALSE}
speeds up the calculation of the dependency model.}
}
\value{
\linkS4class{DependencyModel}
}
\description{
Probabilistic PCA.
}
\examples{
# ppca()
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
A Probabilistic Interpretation of Canonical Correlation Analysis, Bach
Francis R. and Jordan Michael I. 2005 Technical Report 688. Department of
Statistics, University of California, Berkley.

Probabilistic Principal Component Analysis, Tipping Michael E. and Bishop
Christopher M. 1999. \emph{Journal of the Royal Statistical Society}, Series
B, \bold{61}, Part 3, pp. 611--622.

EM Algorithms for ML Factorial Analysis, Rubin D. and Thayer D. 1982.
\emph{Psychometrika}, \bold{vol. 47}, no. 1.
}
\keyword{math}

