% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.phi.R
\name{update_phi_isotropic}
\alias{update_phi_isotropic}
\title{Update Phi (Isotropic)}
\usage{
update_phi_isotropic(Xcov, W, epsilon, dx)
}
\arguments{
\item{Xcov}{Data}

\item{W}{W}

\item{epsilon}{epsilon}

\item{dx}{dx}
}
\value{
numeric
}
\description{
Update Phi (Isotropic).
}
\author{
Olli-Pekka Huovilainen and Leo Lahti \email{leo.lahti@iki.fi}
}

