% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{class}
\name{DependencyModel-class}
\alias{DependencyModel-class}
\alias{getModelMethod}
\alias{getModelMethod,DependencyModel-method}
\alias{getParams}
\alias{getParams,DependencyModel-method}
\alias{getPhi}
\alias{getPhi,DependencyModel-method}
\alias{getScore}
\alias{getScore,DependencyModel-method}
\alias{getW}
\alias{getW,DependencyModel-method}
\alias{getWindowSize}
\alias{getWindowSize,DependencyModel-method}
\alias{getZ}
\alias{getZ,DependencyModel-method}
\title{Class DependencyModel}
\description{
A Dependency model for one or two data sets.
}
\section{Objects from the Class}{
 Returned by
\code{\link{fit.dependency.model}}, \code{\link{ppca}}, \code{\link{pfa}},
and \code{\link{pcca}} functions. % and \code{\link{pcca.isotropic}}
}
\examples{
  #data(modelData) # Load example data X, Y
  #model <- fit.dependency.model(X, Y)
  ## Getting the latent variable Z when it has been calculated with the model
  ## getZ(model)
}
\author{
Olli-Pekka Huovilainen \email{ohuovila@gmail.com}
}
\keyword{classes}

