% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMT.R
\name{DMT}
\alias{DMT}
\title{Embed Drum Machine Test in battery}
\usage{
DMT(num_trials = 5L, tempo = 100, num_example_trials = 3)
}
\arguments{
\item{num_example_trials}{}
}
\description{
Embed Drum Machine Test in battery
}
