% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMT.R
\name{DMT_languages}
\alias{DMT_languages}
\title{DMT languages}
\usage{
DMT_languages()
}
\description{
Lists the languages available for DMT implementations. Muss mit den
Spaltennamen (in Kleinschreibung) in data_raw/DMT_dict.xlsx übereinstimmen.
}
