% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMT.R
\name{DMT_standalone}
\alias{DMT_standalone}
\title{Standalone Drum Machine Test}
\usage{
DMT_standalone(
  tempo = 100,
  num_trials = 5L,
  num_examples = 3,
  with_feedback = TRUE,
  stratified_sampling = TRUE,
  custom_stratified_sampling_allocation = NULL,
  trial_timeout = 90,
  language = "en"
)
}
\arguments{
\item{language}{Scalar character, one of DMT_languages() (currently
"en", "de", "de_f"). Fixes the standalone test to exactly this language.}
}
\description{
Standalone Drum Machine Test
}
