\name{expdata}
\alias{expdata1}
\alias{expdata2}
\docType{data}
\title{ Example data for drCCA }
\description{
            Randomly generated data set for the examples.
            Data matrix with 2000 rows and 6 columns.
  Rows corresponds to the samples and each column corresponds to a
  particular features.
}
\usage{data(expdata1)}
\format{
         A data matrix with 2000 rows and 6 columns. Column names and
   row names are ommitted.
}
\details{
            A randomly generated data set for the example purpose only.
}

\examples{
#data(expdata1)
}
\keyword{datasets}
